==> code/zz_archive/01_getRefNonRef.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.colors as pltc
import geopandas
import dataretrieval.nwis as nwis
import contextily as cx
from matplotlib.ticker import FormatStrFormatter
import matplotlib as mpl
mpl.rcParams['pdf.fonttype'] = 42

dat = pd.read_csv("../data/dryingRegimes_data_RF.csv")
gages = dat.gage.astype("string").str.zfill(width=8)

siteINFO_huc = nwis.get_info(huc='17', siteType='ST')
huc17 = pd.DataFrame(siteINFO_huc[0].site_no).astype(int)
huc17.columns = ['gage']

NP_HUC17 = pd.merge(dat,huc17,on='gage')

NP_summary = NP_HUC17.groupby("gage").size().reset_index(name='counts')

NP_gage = pd.DataFrame(NP_HUC17.gage.unique())
NP_gage = NP_gage[0].astype("string").str.zfill(width=8)

NP_summary.to_csv('../data/PNW_DryingRegimeGages.csv')

npPNW = pd.read_csv('../data/PNW_DryingRegimeGages.csv')

classDat = pd.read_csv('../data/mean_annual_no_flow_and_climate_metric_means_for_no_flow_sites_082819_with_info.csv')
classDat = classDat[['site', 'CLASS']]
classDat.columns = ['gage', 'CLASS']

NP = pd.merge(npPNW,classDat,on='gage')

NP.to_csv('../data/PNW_DryingRegimeGages.csv')
