==> code/zz_archive/03_getVIC_PRMS_data.ipynb <==
import netCDF4 as nc
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import geopandas
import xarray as xr
import matplotlib.colors as pltc
import geopandas
import dataretrieval.nwis as nwis
import contextily as cx

## Cubic feet to cubic meters conversion factor
cfs_2_cms = 0.0283168466


#### Load NP sites from Drying Regimes

pnwNP = pd.read_csv("../data/pnwNP_Info.csv")

#### Cross-reference with CIG provided shapefile to get stream segment

shp = geopandas.read_file("../data/VIC_UW/shapefiles/columbia_seg.shp")
shp = shp.loc[shp['POI_ID'].isin(pnwNP["comid"])]
seg_ids = np.array(shp['seg_id'])

#### Get VIC data for corresponding stream segments and write to new netCDF.

## Open modeled datasets
data = xr.open_mfdataset('../data/VIC_UW/vic_historical_first_route_all.nc')

## Extract data at segment locations.
data = data.where(data['reachID'].isin(seg_ids),drop=True)

# # sort seg_ids in ascending order (not necessary)
# data['seg'] = sorted(seg_ids)

# # make new output directory
# Path(f'/home/disk/rocinante/rawrgers/Data/Data_Requests/Rupp/nwfsc_data/{hydro}/projections/first_route/{model}/{scenario}/').mkdir(parents=True, exist_ok=True)


# # # save the reachID (i.e., seg_ids) and IRFroutedRunoff (i.e., streamflow) to new netcdf file
# gage = '13152500'
# name = '../data/VIC_UW/' + gage
name = "../data/pnwNP_Gages_VIC.nc"

# out = data[['reachID','IRFroutedRunoff']].to_netcdf(name)

#### Get PRMS data for corresponding stream segments and write to new netCDF.

## Open modeled datasets
data = xr.open_mfdataset('../data/VIC_UW/prms_historical_first_route_all.nc')

## Extract data at segment locations.
data = data.where(data['reachID'].isin(seg_ids),drop=True)

# # sort seg_ids in ascending order (not necessary)
# data['seg'] = sorted(seg_ids)

# # make new output directory
# Path(f'/home/disk/rocinante/rawrgers/Data/Data_Requests/Rupp/nwfsc_data/{hydro}/projections/first_route/{model}/{scenario}/').mkdir(parents=True, exist_ok=True)


# # # save the reachID (i.e., seg_ids) and IRFroutedRunoff (i.e., streamflow) to new netcdf file
# gage = '13152500'
# name = '../data/VIC_UW/' + gage
name = "../data/pnwNP_Gages_PRMS.nc"

# out = data[['reachID','IRFroutedRunoff']].to_netcdf(name)

## Example for Matt

seg_ids = [909]

## Open modeled datasets
data = xr.open_mfdataset('../data/VIC_UW/vic_historical_first_route_all.nc')

## Extract data at segment locations.
data = data.where(data['reachID'].isin(seg_ids),drop=True)

dat = data[['reachID','IRFroutedRunoff']]
comb = dat.to_dataframe()
comb

maladVIC = pd.read_csv('../data/UW_StreamflowOutput/historical_livneh_VIC_P1-WOODGOOD-streamflow-1.0.csv',comment='#')
maladVIC['date'] = pd.to_datetime(maladVIC['date'])
maladVIC.columns = ['time','streamflowVIC']
maladVIC['streamflowVIC'] = maladVIC['streamflowVIC'] * cfs_2_cms
maladVIC

## Scratch space
