==> code/zz_archive/05_pnwModelData_NWM2d1.ipynb <==
import multiprocessing
import numpy as np
import multiprocess as mp
import os
import fsspec
import pandas as pd
import xarray as xr
from tqdm import tqdm
import geopandas
import pandas as pd
import datetime as dt
import time as time

pnwNP = pd.read_csv("../data/pnwNP_Info.csv")

### Pull out sites
shp = geopandas.read_file("../data/VIC_UW/shapefiles/columbia_seg.shp")
shp = pd.merge(pnwNP,shp,left_on='comid',right_on='POI_ID').fillna(0)

fs = fsspec.filesystem('s3', anon=True)
_file = fs.glob('noaa-nwm-retrospective-2-1-zarr-pds/chrtout.zarr')

ds = xr.open_dataset(fs.get_mapper(_file[0]), engine='zarr', backend_kwargs={'consolidated': True})

for i in range(len(shp)):
    print(shp['comid'][i])
    ## Get NWM 2.1 data
    # slice all data using a specific reach identifier and time range
    dat = pd.read_csv('../data/pnwNP_modeledData/'+str(shp['gage'][i])+".csv")
    dat['time'] = pd.to_datetime(dat['time'])
    timerange = slice(dat['time'][0], dat['time'].iloc[-1])
    df = ds.sel(feature_id=shp['comid'][i],
                 time=timerange).streamflow.persist() 

    NWM = pd.DataFrame(df.to_pandas()).resample('1d').mean()
    NWM['time'] = pd.to_datetime(NWM.index)
    NWM.columns = ["streamflow_NWM2d1","time"]
    NWM = NWM.reset_index(drop=True)
    NWM["gage"] = shp['gage'][i]



    ## Combine with previous modeled data
    datMain = pd.merge(dat,NWM, on='time',how='inner')

    datMain.to_csv('../data/pnwNP_modeledData/'+str(shp['gage'][i])+"wNWM2d1.csv")

def NWMmp(i):
    print(shp['comid'][i])
    ## Get NWM 2.1 data
    # slice all data using a specific reach identifier and time range
    dat = pd.read_csv('../data/pnwNP_modeledData/'+str(shp['gage'][i])+".csv")
    dat['time'] = pd.to_datetime(dat['time'])
    timerange = slice(dat['time'][0], dat['time'].iloc[-1])
    df = ds.sel(feature_id=shp['comid'][i],
                 time=timerange).streamflow.persist() 

    NWM = pd.DataFrame(df.to_pandas()).resample('1d').mean()
    NWM['time'] = pd.to_datetime(NWM.index)
    NWM.columns = ["streamflow_NWM2d1","time"]
    NWM = NWM.reset_index(drop=True)
    NWM["gage"] = shp['gage'][i]



    ## Combine with previous modeled data
    datMain = pd.merge(dat,NWM, on='time',how='inner')

    datMain.to_csv('../data/pnwNP_modeledData/'+str(shp['gage'][i])+"wNWM2d1.csv")
    datMain.to_csv('../data/pnwNP_modeledData/'+str(shp['gage'][i])+"wNWM2d1.csv")
    print(time.process_time())



import numpy as np
import multiprocess as mp
import os


if __name__ == '__main__':
    N = mp.cpu_count()

    with mp.Pool(processes = N) as p:
        p.map(NWMmp, list(range(len(shp))))
        p.close()

# This step takes a bit longer because it's actually returning the data
dat = dat.resample(time='1d').mean()

NWM = pd.DataFrame(dat.to_pandas())
NWM['time'] = pd.to_datetime(NWM.index)
NWM.columns = ["streamflow_NWM2d1","time"]
NWM = NWM.reset_index(drop=True)
NWM["gage"] = shp['gage'][i]


! pip3 install multiprocess

import numpy as np
import multiprocess as mp
import os


if __name__ == '__main__':
    N= mp.cpu_count()
    
    def NWM(i):
        print(shp['comid'][i])
    #     ## Get NWM 2.1 data
    #     # slice all data using a specific reach identifier and time range
    #     dat = pd.read_csv('../data/pnwNP_modeledData/'+str(shp['gage'][i])+".csv")
    #     dat['time'] = pd.to_datetime(dat['time'])
    #     timerange = slice(dat['time'][0], dat['time'].iloc[-1])
    #     dat = ds.sel(feature_id=shp['comid'][i],
    #                  time=timerange).streamflow.persist() 

    #     NWM = pd.DataFrame(dat.to_pandas()).resample('1d').mean()
    #     NWM['time'] = pd.to_datetime(NWM.index)
    #     NWM.columns = ["streamflow_NWM2d1","time"]
    #     NWM = NWM.reset_index(drop=True)
    #     NWM["gage"] = shp['gage'][i]



    #     ## Combine with previous modeled data
    #     datMain = pd.merge(dat,NWM, on='time',how='inner')
        datMain = pd.DataFrame([0,1,2])
        datMain.to_csv('../data/pnwNP_modeledData/'+str(shp['gage'][i])+"wNWM2d1.csv")

    with mp.Pool(processes = N) as p:
        p.map(NWM, list(range(len(shp))))

## Combine with previous modeled data
datMain = pd.merge(dat,NWM, on='time',how='inner')

datMain.to_csv('../data/pnwNP_modeledData/'+str(shp['gage'][i])+"wNWM2d1.csv")

# timerange = slice('2011-01-31', '2011-02-28')
# dat = ds.sel(feature_id=comid,time=timerange).streamflow.persist() 
dat = ds.sel(feature_id=comid).streamflow.persist() 

dat = dat.resample(time='1d').mean()

tetNWM = pd.DataFrame(dat.to_pandas())

NWM = tetNWM

NWM = pd.DataFrame(dat.to_pandas())
NWM['time'] = pd.to_datetime(NWM.index)
NWM.columns = ["streamflow_NWM2d1","time"]
NWM = NWM.reset_index(drop=True)
# NWM["gage"] = shp['gage'][i]
NWM
