==> code/zz_archive/07_ProsperAlign.ipynb <==
import netCDF4 as nc
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import geopandas
import xarray as xr
import matplotlib.colors as pltc
import geopandas
import datetime as dt
from scipy import stats
from sklearn import preprocessing
import s3fs
import matplotlib as mpl
import rasterio
from rasterio.crs import CRS
import rioxarray as rxr
import rasterstats as rs
mpl.rcParams['pdf.fonttype'] = 42

## Cubic feet to cubic meters conversion factor
cfs_2_cms = 0.0283168466

src = rasterio.open("../data/PROSPER/SPP_MEAN.tif")
srcArry = src.read(1)
srcArry[srcArry < 0] = np.nan ## Filter out 


pnwNP = pd.read_csv("../data/pnwNP_StatsLocations.csv")
pnwNP = geopandas.GeoDataFrame(pnwNP, geometry=geopandas.points_from_xy(pnwNP.dec_long_va,pnwNP.dec_lat_va)) \
.set_crs('epsg:4269',inplace=True)

pnwNP = pnwNP.to_crs(src.crs)
# pnwNP = pnwNP.iloc[2:4]
pnwNP['geometry'] = pnwNP.geometry.buffer(100)
pnwNP.to_file('../data/pnwBufferTestDelete.shp')

tt = rs.zonal_stats("../data/pnwBufferTestDelete.shp",
                    srcArry,
                    nodata=-999,
                    affine=src.transform,
                    geojson_out=True,
                    copy_properties=True,
                    stats="count min mean max median")

out = geopandas.GeoDataFrame.from_features(tt)

out[out['count'] == out['count'].max()]

out

out = out.drop(['geometry','gage'],axis=1)
out.columns = ['gage', 'dec_lat', 'dec_long', 'numZero',
       'numOne', 'startDate', 'endDate', 'comid', 'numObs', 'propZero',
       'propOne', 'minProb_PROSPER', 'maxProb_PROSPER', 'meanProb_PROSPER', 'countProb_PROSPER', 'medianProb_PROSPER']

out.to_csv("../data/pnwNP_StatsLocations_SS.csv")

out[out['comid'].isna()]
